On FreeBSD, use the expat package, if available, in preference to base system xml support.

+++ csrc/conf/FreeBSD.sh
@@ -15,7 +15,8 @@ echo "# FreeBSD.sh $@ was here" >> conf.mk
 set -x
 mkdir -p include
 find include lib -type l -name \*expat\* -exec rm '{}' ';'
-# XXX fix this - should check for version installed by ports and use that
+test -f /usr/local/include/expat.h && exit 0
+echo "# Using bsdxml.h for expat - recommend installing expat package and repeat ./configure" >> conf.mk
 # Make some symlinks so that we can use the native expat.
 ln -s /usr/include/bsdxml.h include/expat.h
 ln -s /usr/lib/libbsdxml.a lib/libexpat.a
